Address PR review feedback: fix event delivery, flush, types, and heuristic
- Fix critical bug: runtime events arriving before handler registration were
  silently dropped. Added deliver() helper using the same buffer-or-dispatch
  pattern as the existing onmessage flow.
- Flush composed FanOutEventSink on shutdown instead of only the file sink.
- Add warning when WASM worker falls back to activeCallIds heuristic for callId.
- Fix UTF-8 vs UTF-16 mismatch in isLikelyVariable decoration heuristic by
  using TextEncoder for byte-accurate comparison.
- Add | null to DeserializedEventKind fields to match deserializer output.
- Extract deserializeRuntimeEvent to shared module and update WASM worker to
  deserialize events and use formatValue instead of removed formatValueShort.
